+++ tests/Settings/ModuleTest.php
@@ -4,7 +4,7 @@ namespace Svycka\SettingsTest;
 
 use Svycka\Settings\Module;
 
-class ModuleTest extends \PHPUnit_Framework_TestCase
+class ModuleTest extends \PHPUnit\Framework\TestCase
 {
     public function testConfigIsArray()
     {