+++ tests/Settings/Service/Factory/SettingsServiceFactoryTest.php
@@ -6,13 +6,13 @@ use Interop\Container\ContainerInterface;
 use Svycka\Settings\Collection\CollectionsManager;
 use Svycka\Settings\Service\Factory\SettingsServiceFactory;
 use Svycka\Settings\Service\SettingsService;
-use Zend\ServiceManager\ServiceManager;
+use Laminas\ServiceManager\ServiceManager;
 
 /**
  * @author Vytautas Stankus <svycka@gmail.com>
  * @license MIT
  */
-class SettingsServiceFactoryTest extends \PHPUnit_Framework_TestCase
+class SettingsServiceFactoryTest extends \PHPUnit\Framework\TestCase
 {
     public function testCanCreate()
     {