+++ tests/Settings/Type/Factory/TypesManagerFactoryTest.php
@@ -5,13 +5,13 @@ namespace Svycka\SettingsTest\Type\Factory;
 use Svycka\Settings\Options\ModuleOptions;
 use Svycka\Settings\Type\Factory\TypesManagerFactory;
 use Svycka\Settings\Type\TypesManager;
-use Zend\ServiceManager\ServiceLocatorInterface;
+use Laminas\ServiceManager\ServiceLocatorInterface;
 
 /**
  * @author Vytautas Stankus <svycka@gmail.com>
  * @license MIT
  */
-class TypesManagerFactoryTest extends \PHPUnit_Framework_TestCase
+class TypesManagerFactoryTest extends \PHPUnit\Framework\TestCase
 {
     public function testCanCreate()
     {